Transactions/fastapi_backend/app/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
pytesseract==0.3.10
opencv-python==4.9.0.80
numpy==1.26.4
pymongo==4.6.3
google-generativeai==0.5.4
